+++ officialplugins/ReCaptcha/src/main/java/dev/truewinter/snowmail/officialplugins/recaptcha/Listener.java
@@ -1,4 +1,4 @@
-package dev.truewinter.snowmail.officialplugins.snowcaptcha;
+package dev.truewinter.snowmail.officialplugins.recaptcha;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -12,9 +12,9 @@ import dev.truewinter.snowmail.api.pojo.objects.Form;
 import java.io.OutputStream;
 import java.net.HttpURLConnection;
 import java.net.URL;
+import java.nio.charset.StandardCharsets;
 import java.util.HashMap;
 
-@SuppressWarnings("unused")
 public class Listener implements dev.truewinter.PluginManager.Listener {
     private final Logger logger;
 
@@ -24,43 +24,39 @@ public class Listener implements dev.truewinter.PluginManager.Listener {
 
     @EventHandler
     public void onFormSubmission(FormSubmissionEvent e) {
-        if (!Form.recursivelyGetInputs(e.getForm().getInputs()).containsKey("snowcaptcha")) return;
+        if (!Form.recursivelyGetInputs(e.getForm().getInputs()).containsKey("g-recaptcha-response")) return;
 
         HashMap<String, String> metadata = e.getForm().getMetadata();
-        if (!metadata.containsKey("snowcaptcha-sitekey") || !metadata.containsKey("snowcaptcha-secret") || !metadata.containsKey("snowcaptcha-host")) {
-            logger.warn("Metadata keys snowcaptcha-sitekey, snowcaptcha-secret, and snowcaptcha-host must be set, skipping captcha validation");
+        if (!metadata.containsKey("recaptcha-sitekey") || !metadata.containsKey("recaptcha-secret")) {
+            logger.warn("Metadata keys recaptcha-sitekey, and recaptcha-secret must be set, skipping captcha validation");
             return;
         }
 
-        FormSubmissionInput token = e.getSubmission().get("snowcaptcha");
+        FormSubmissionInput token = e.getSubmission().get("g-recaptcha-response");
         if (token == null || Util.isBlank(token.value())) {
             e.setError("Captcha is required");
             return;
         }
 
         try {
-            URL url = new URL(getValidateURL(metadata.get("snowcaptcha-host")));
+            URL url = new URL("https://www.google.com/recaptcha/api/siteverify");
             HttpURLConnection connection = (HttpURLConnection) url.openConnection();
             connection.setRequestMethod("POST");
             connection.setDoOutput(true);
-            connection.setRequestProperty("Content-Type", "application/json");
+            connection.setRequestProperty("Content-Type", "application/x-www-form-urlencoded");
             connection.setRequestProperty("Accept", "application/json");
             connection.setConnectTimeout(15000);
             connection.setReadTimeout(15000);
 
-            HashMap<String, String> body = new HashMap<>();
-            body.put("token", token.value());
-            body.put("sitekey", metadata.get("snowcaptcha-sitekey"));
-            body.put("secretkey", metadata.get("snowcaptcha-secret"));
-
+            String body = String.format("secret=%s&response=%s", metadata.get("recaptcha-secret"), token.value());
             connection.connect();
             try (OutputStream outputStream = connection.getOutputStream()) {
-                outputStream.write(new ObjectMapper().writeValueAsBytes(body));
+                outputStream.write(body.getBytes(StandardCharsets.UTF_8));
             }
 
             HashMap<String, Object> response = new ObjectMapper()
                     .readValue(connection.getInputStream().readAllBytes(), new TypeReference<>() {});
-            boolean valid = (boolean) response.get("valid");
+            boolean valid = (boolean) response.get("success");
             if (!valid) {
                 e.setError("Invalid captcha token");
             }
@@ -70,15 +66,5 @@ public class Listener implements dev.truewinter.PluginManager.Listener {
             logger.error("Failed to validate captcha token", ex);
             e.setError("Failed to validate captcha");
         }
-    }
-
-    private String getValidateURL(String host) {
-        StringBuilder stringBuilder = new StringBuilder();
-        stringBuilder.append(host);
-        if (!host.endsWith("/")) {
-            stringBuilder.append("/");
-        }
-        stringBuilder.append("validate-token");
-        return stringBuilder.toString();
     }
 }